requirements.txt
fastapi
uvicorn
openai
bs4
requests
python-dotenv
faiss-cpu           # or faiss-gpu if needed
tiktoken
pydantic
sentence-transformers
PyJWT
streamlit
gradio
httpx
python-jose 